Implement 'trackPage' in MixPanel adapter
As Mixpanel is mostly centered around Events as opposed to Page Views,
it doesn't strictly have an API for tracking page views. This is the
reason that the trackPage function was not initially implemented. However, it is
useful to track page views as events.

The implementation for the trackPage function is as suggested by the
[MixPanel
docs](https://mixpanel.com/help/questions/articles/why-measure-events-vs-page-views)