Update path view tracking example
I was setting this thing up today and wanted to track page views.  I happen to be hosting my Ember app on Github Pages so I need to use a hash-based URL.  The example code here reported every URL the same way, so I dug around and found that `this.get("url")` reported a value that actually reflects the real URL, regardless of the location type.