Refactor canUseMetrics
By moving the check to the service layer we can ensure user defined
adapters, and all future adapters, take advantage of this feature and
that we needn't litter the code with this check.

Note that doNotTrack is a deprecated feature, and thus checking to see
if a user has opted in isn't sufficient -- we also have to ensure the
absence of this feature (should it be removed) doesn't break our code
as well.