[refactor] remove `object-transforms#isPresent`
maintaining this utility function is unnecessary since we already have `isPresent` from `@ember/utils`, and also confusing because we were doing some name-alias hoop-jumping to use it, and also using `@ember#isPresent` and `object-transforms#isPresent` in different contexts throughout the addon for the same purpose. we should align on Ember's solution here.

- removes `object-transforms#isPresent`
  - revert to bare import name for internal `object-transforms` usage
- refactor various imports to `@ember/utils#isPresent`