[refactor] `object-transforms#isPresent` -> `hasOwnProperties`
current `isPresent` function is name-aliasing `@ember-utils#isPresent`, but serves a different functional purpose. Ember's `isPresent` will return `true` for an object with no keys, but the the util function we're exposing as `isPresent` is checking if the given object has properties. given that this is an Ember addon, we should avoid overloading the concept of "present" in the context of an object's property names.

- renames `object-transforms#isPresent` to `object-transforms#hasOwnProperties`
- updates imports and instances throughout
- collapses import of `@ember/utils#isPresent` in `object-transforms#compact`